One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/dispatch/route-info/route-info.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { RouteInfoComponent } from './route-info.component';

describe('Component: RouteInfo', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [RouteInfoComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([RouteInfoComponent],
      (component: RouteInfoComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(RouteInfoComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(RouteInfoComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-route-info></app-route-info>
  `,
  directives: [RouteInfoComponent]
})
class RouteInfoComponentTestController {
}

